iOS support (#801)
* Modifications to support iOS compilation

* Use static_cast and remove .DS_Store file

* Add .DS_store to git ignore

+++ include/cpr/body.h
@@ -31,7 +31,7 @@ class Body : public StringHolder<Body> {
         std::ifstream is(file.filepath, std::ifstream::binary);
         if (is) {
             is.seekg(0, is.end);
-            int length = is.tellg();
+            int length = static_cast<int>(is.tellg());
             is.seekg(0, is.beg);
             std::vector<char> buffer;
             buffer.resize(length);
